test_api.py
from google import genai
import sys
import os
from dotenv import load_dotenv

load_dotenv()

try:
    client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))
    for m in client.models.list():
        print(m.name)
except Exception as e:
    print(f"Error: {e}")
