requirements.txt
streamlit
pandas
supabase
google-genai
python-dotenv
